+++ TelegramBotAPI-extensions-utils/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/utils/updates/retrieving/LongPolling.kt
@@ -1,10 +1,10 @@
-package com.github.insanusmokrassar.TelegramBotAPI.extensions.api.updates
+package com.github.insanusmokrassar.TelegramBotAPI.extensions.utils.updates.retrieving
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.bot.exceptions.RequestException
-import com.github.insanusmokrassar.TelegramBotAPI.extensions.api.InternalUtils.convertWithMediaGroupUpdates
-import com.github.insanusmokrassar.TelegramBotAPI.extensions.api.InternalUtils.lastUpdateIdentifier
-import com.github.insanusmokrassar.TelegramBotAPI.extensions.api.getUpdates
+import com.github.insanusmokrassar.TelegramBotAPI.extensions.utils.updates.convertWithMediaGroupUpdates
+import com.github.insanusmokrassar.TelegramBotAPI.extensions.utils.updates.lastUpdateIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.requests.GetUpdates
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdates.*
@@ -13,8 +13,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.updateshandlers.*
 import com.github.insanusmokrassar.TelegramBotAPI.utils.*
 import kotlinx.coroutines.*
 
-@Deprecated("Replaced and renamed in TelegramBotAPI-extensions-utils")
-fun RequestsExecutor.startGettingOfUpdates(
+fun RequestsExecutor.startGettingOfUpdatesByLongPolling(
     timeoutSeconds: Seconds = 30,
     scope: CoroutineScope = CoroutineScope(Dispatchers.Default),
     exceptionsHandler: (ExceptionHandler<Unit>)? = null,
@@ -32,10 +31,12 @@ fun RequestsExecutor.startGettingOfUpdates(
                 }
             }
         ) {
-            val updates = getUpdates(
-                offset = lastUpdateIdentifier?.plus(1),
-                timeout = timeoutSeconds,
-                allowed_updates = allowedUpdates
+            val updates = execute(
+                GetUpdates(
+                    offset = lastUpdateIdentifier?.plus(1),
+                    timeout = timeoutSeconds,
+                    allowed_updates = allowedUpdates
+                )
             ).let { originalUpdates ->
                 val converted = originalUpdates.convertWithMediaGroupUpdates()
                 /**
@@ -71,25 +72,23 @@ fun RequestsExecutor.startGettingOfUpdates(
 @FlowPreview
 @PreviewFeature
 @Suppress("unused")
-@Deprecated("Replaced and renamed in TelegramBotAPI-extensions-utils")
-fun RequestsExecutor.startGettingFlowsUpdates(
+fun RequestsExecutor.startGettingFlowsUpdatesByLongPolling(
     timeoutSeconds: Seconds = 30,
     scope: CoroutineScope = CoroutineScope(Dispatchers.Default),
-    exceptionsHandler: (suspend (Exception) -> Unit)? = null,
+    exceptionsHandler: ExceptionHandler<Unit>? = null,
     flowsUpdatesFilterUpdatesKeeperCount: Int = 64,
     flowUpdatesPreset: FlowsUpdatesFilter.() -> Unit = {}
 ): FlowsUpdatesFilter = FlowsUpdatesFilter(flowsUpdatesFilterUpdatesKeeperCount).apply {
     flowUpdatesPreset()
-    startGettingOfUpdates(timeoutSeconds, scope, exceptionsHandler, allowedUpdates, asUpdateReceiver)
+    startGettingOfUpdatesByLongPolling(timeoutSeconds, scope, exceptionsHandler, allowedUpdates, asUpdateReceiver)
 }
 
-@Deprecated("Replaced and renamed in TelegramBotAPI-extensions-utils")
-fun RequestsExecutor.startGettingOfUpdates(
+fun RequestsExecutor.startGettingOfUpdatesByLongPolling(
     updatesFilter: UpdatesFilter,
     timeoutSeconds: Seconds = 30,
-    exceptionsHandler: (suspend (Exception) -> Unit)? = null,
+    exceptionsHandler: ExceptionHandler<Unit>? = null,
     scope: CoroutineScope = CoroutineScope(Dispatchers.Default)
-): Job = startGettingOfUpdates(
+): Job = startGettingOfUpdatesByLongPolling(
     timeoutSeconds,
     scope,
     exceptionsHandler,
@@ -97,8 +96,7 @@ fun RequestsExecutor.startGettingOfUpdates(
     updatesFilter.asUpdateReceiver
 )
 
-@Deprecated("Replaced and renamed in TelegramBotAPI-extensions-utils")
-fun RequestsExecutor.startGettingOfUpdates(
+fun RequestsExecutor.startGettingOfUpdatesByLongPolling(
     messageCallback: UpdateReceiver<MessageUpdate>? = null,
     messageMediaGroupCallback: UpdateReceiver<MessageMediaGroupUpdate>? = null,
     editedMessageCallback: UpdateReceiver<EditMessageUpdate>? = null,
@@ -115,10 +113,10 @@ fun RequestsExecutor.startGettingOfUpdates(
     pollCallback: UpdateReceiver<PollUpdate>? = null,
     pollAnswerCallback: UpdateReceiver<PollAnswerUpdate>? = null,
     timeoutSeconds: Seconds = 30,
-    exceptionsHandler: (suspend (Exception) -> Unit)? = null,
+    exceptionsHandler: ExceptionHandler<Unit>? = null,
     scope: CoroutineScope = GlobalScope
 ): Job {
-    return startGettingOfUpdates(
+    return startGettingOfUpdatesByLongPolling(
         SimpleUpdatesFilter(
             messageCallback,
             messageMediaGroupCallback,
@@ -143,8 +141,7 @@ fun RequestsExecutor.startGettingOfUpdates(
 }
 
 @Suppress("unused")
-@Deprecated("Replaced and renamed in TelegramBotAPI-extensions-utils")
-fun RequestsExecutor.startGettingOfUpdates(
+fun RequestsExecutor.startGettingOfUpdatesByLongPolling(
     messageCallback: UpdateReceiver<MessageUpdate>? = null,
     mediaGroupCallback: UpdateReceiver<MediaGroupUpdate>? = null,
     editedMessageCallback: UpdateReceiver<EditMessageUpdate>? = null,
@@ -158,9 +155,9 @@ fun RequestsExecutor.startGettingOfUpdates(
     pollCallback: UpdateReceiver<PollUpdate>? = null,
     pollAnswerCallback: UpdateReceiver<PollAnswerUpdate>? = null,
     timeoutSeconds: Seconds = 30,
-    exceptionsHandler: (suspend (Exception) -> Unit)? = null,
+    exceptionsHandler: ExceptionHandler<Unit>? = null,
     scope: CoroutineScope = CoroutineScope(Dispatchers.Default)
-): Job = startGettingOfUpdates(
+): Job = startGettingOfUpdatesByLongPolling(
     messageCallback = messageCallback,
     messageMediaGroupCallback = mediaGroupCallback,
     editedMessageCallback = editedMessageCallback,
@@ -180,3 +177,4 @@ fun RequestsExecutor.startGettingOfUpdates(
     exceptionsHandler = exceptionsHandler,
     scope = scope
 )
+